+++ customerservice/src/main/java/com/etiya/customerservice/dto/address/ListAddressResponseDto.java
@@ -4,13 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 public class ListAddressResponseDto {
 
-            private Integer customerId;
-            private Integer neighbourhoodId;
+            private Long customerId;
+            private Long neighbourhoodId;
             private String addressName;
             private String street;
             private Integer houseNumber;
