+++ customerservice/src/main/java/com/etiya/customerservice/dto/contactinformation/ListContactInformationResponseDto.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 public class ListContactInformationResponseDto {
-    private Integer customerId;
+    private Long customerId;
     private String contactName;
     private String email;
     private Integer homePhone;
