+++ customerservice/src/main/java/com/etiya/customerservice/dto/neighbourhood/CreateNeighbourhoodRequestDto.java
@@ -15,5 +15,5 @@ public class CreateNeighbourhoodRequestDto {
     private String name;
     @NotNull
     @NotBlank
-    private Integer districtId;
+    private Long districtId;
 }
